Keep the template catalogue out of the gallery's bundle
A gallery card shows a name, a tagline, tags, a palette and two numbers.
Importing the catalogue from a client component shipped every template's full
section copy with it, which is the bulk of the file and none of this page.

Reading it in a server shell and passing down only what a card renders, the
same shape the landing page already uses.

Measured over the wire against a production build, JS bytes with the cache
cleared, identical across three runs:

  /templates  476.8 KiB -> 466.4 KiB
  /           353.2 KiB -> 342.9 KiB

The landing page moves too because the catalogue was in a chunk shared across
client entries. It is still reachable from the editor and the preview route,
which need the sections, so this does not remove it from the graph.
